+++ R/other.R
@@ -252,3 +252,31 @@ insert <- function(file, after, what){
   writeLines(new, con = con)
   close(con)
 }
+
+#' Make Progress Bar
+#'
+#' @param i The current iteration.
+#' @param k Total iterations.
+#' @param numTicks The result of \code{progress}.
+#' @return The next \code{numTicks} argument.
+#'
+#' @export
+progress <- function(i, k, numTicks){
+
+  if(i == 1) numTicks <- 0
+
+  if(numTicks == 0) cat("|-")
+
+  while(i > numTicks*(k/40)){
+
+    cat("-")
+    if(numTicks == 10) cat("(25%)")
+    if(numTicks == 20) cat("(50%)")
+    if(numTicks == 30) cat("(75%)")
+    numTicks <- numTicks + 1
+  }
+
+  if(i == k) cat("-|\n")
+
+  return(numTicks)
+}