+++ NEWS.md
@@ -3,6 +3,7 @@
 * Have `migraph` functions coerce arguments as character input
 * Fix errors in documentation and package imports
 * Fix `simpliGSEA` pruning so it actually works
+* Add code for `progress` bar function
 
 ## miSciTools 1.0.1
 ---------------------